Switch to esbuild (#199)
* Switch to esbuild

* built => dist for 'types' field

* @types/mocha => @types/jest

* Don't use destructuring on this.coords

* Fix #203: Copy underlying data in c-contiguous order

* Fix naming

* New dist

* Fix CI

* Update dependencies

* libpython2-dev => python-dev

* Simplify toArray

* Cleanup

* Update deps

* Add benchmarks

* Update license year

* Add .npmignore

* Publish coverage to CodeClimate

* Layout update

* Fix CI

* Update badges

+++ src/core/reshape.spec.ts
@@ -8,15 +8,15 @@ import { array } from './array';
 
 describe('(v) reshape', () => {
   it('should throw error if new shape does not match length', () => {
-    const f32 = new Float64Array([1, 2, 3, 4]);
-    const x = array(f32);
+    const f64 = new Float64Array([1, 2, 3, 4]);
+    const x = array(f64);
 
     throws(() => { x.reshape(1, 2); }, Error);
   });
 
   it('should be able to create row vector of column vector', () => {
-    const f32 = new Float64Array([1, 2, 3, 4]);
-    const x = array(f32, { shape: [1, 4] });
+    const f64 = new Float64Array([1, 2, 3, 4]);
+    const x = array(f64, { shape: [1, 4] });
 
     equals(x.shape, [1, 4]);
     x.reshape(4, 1);
@@ -24,8 +24,8 @@ describe('(v) reshape', () => {
   });
 
   it('should work as the static equivalent', () => {
-    const f32 = new Float64Array([1, 2, 3, 4]);
-    const x = array(f32, { shape: [1, 4] });
+    const f64 = new Float64Array([1, 2, 3, 4]);
+    const x = array(f64, { shape: [1, 4] });
 
     equals(x.shape, [1, 4]);
     equals(reshape(x, 4, 1).shape, [4, 1]);
